+++ src/types/query.d.ts
@@ -54,7 +54,7 @@ export module query {
   export function Identify(ref: ExprArg, password: ExprArg): Expr;
 
   export function Concat(strings: ExprArg, separator?: ExprArg): Expr;
-  export function Casefold(string: ExprArg): Expr;
+  export function Casefold(string: ExprArg, normalizer?: ExprArg): Expr;
 
   export function Time(string: ExprArg): Expr;
   export function Epoch(number: ExprArg, unit: ExprArg): Expr;